agents/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "skillix-backend"
version = "0.1.0"
authors = [
    { name = "Luca Mazzarello and Federico Berman", email = "lucamazza02@gmail.com" }
]
description = "A personalized learning journey creator"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "httpx",
    "python-dotenv",
    "openai",
    "firebase-admin",
    "google-adk",
    "google-cloud-storage",
    "google-cloud-firestore",
    "google-auth",
    "sqlalchemy",
    "PyYAML"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["api", "skillix_agents"] 